Add verbose option and function to find files in the directory.
There are now 2 ways of finding a file in the filesystem: regular
find and scavenge. Regular find traverses the directory structure
in search of the file, while scavenge scans the disk for the file.
The scavenge method relies on the hints on the leader pages and
therefore might not be as accurate (if the filesystem is not
corrupted, it should correct results).